requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.15.2
drf-nested-routers==0.94.1
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
